+++ Assignment1-VSCode/Assignment1-VSCode/src/SearchMethod.java
@@ -60,7 +60,7 @@ public abstract class SearchMethod
         //finds the smallest node in the frontier with the smallest FCost, and pops it out of the frontier. This is used by all heuristic search methods.
         protected MazeState popSmallestFCostFromFrontier()
         {
-            int lowestFCost = FrontierNodes.getFirst().FCost;
+            double lowestFCost = FrontierNodes.getFirst().FCost;
             int lowestIndex = 0;
             
             for(int i = 1; i < FrontierNodes.size(); i++)
